requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
cryptography==37.0.4
h11==0.13.0
idna==3.3
loguru==0.6.0
MouseInfo==0.1.3
outcome==1.2.0
Pillow==9.2.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc==8.5
pyobjc-core==8.5
pyobjc-framework-Accounts==8.5
pyobjc-framework-AddressBook==8.5
pyobjc-framework-AdSupport==8.5
pyobjc-framework-AppleScriptKit==8.5
pyobjc-framework-AppleScriptObjC==8.5
pyobjc-framework-ApplicationServices==8.5
pyobjc-framework-AudioVideoBridging==8.5
pyobjc-framework-AuthenticationServices==8.5
pyobjc-framework-AutomaticAssessmentConfiguration==8.5
pyobjc-framework-Automator==8.5
pyobjc-framework-AVFoundation==8.5
pyobjc-framework-AVKit==8.5
pyobjc-framework-BusinessChat==8.5
pyobjc-framework-CalendarStore==8.5
pyobjc-framework-CFNetwork==8.5
pyobjc-framework-CloudKit==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-Collaboration==8.5
pyobjc-framework-ColorSync==8.5
pyobjc-framework-Contacts==8.5
pyobjc-framework-ContactsUI==8.5
pyobjc-framework-CoreAudio==8.5
pyobjc-framework-CoreAudioKit==8.5
pyobjc-framework-CoreBluetooth==8.5
pyobjc-framework-CoreData==8.5
pyobjc-framework-CoreHaptics==8.5
pyobjc-framework-CoreLocation==8.5
pyobjc-framework-CoreMedia==8.5
pyobjc-framework-CoreMediaIO==8.5
pyobjc-framework-CoreMIDI==8.5
pyobjc-framework-CoreML==8.5
pyobjc-framework-CoreMotion==8.5
pyobjc-framework-CoreServices==8.5
pyobjc-framework-CoreSpotlight==8.5
pyobjc-framework-CoreText==8.5
pyobjc-framework-CoreWLAN==8.5
pyobjc-framework-CryptoTokenKit==8.5
pyobjc-framework-DeviceCheck==8.5
pyobjc-framework-DictionaryServices==8.5
pyobjc-framework-DiscRecording==8.5
pyobjc-framework-DiscRecordingUI==8.5
pyobjc-framework-DiskArbitration==8.5
pyobjc-framework-DVDPlayback==8.5
pyobjc-framework-EventKit==8.5
pyobjc-framework-ExceptionHandling==8.5
pyobjc-framework-ExecutionPolicy==8.5
pyobjc-framework-ExternalAccessory==8.5
pyobjc-framework-FileProvider==8.5
pyobjc-framework-FileProviderUI==8.5
pyobjc-framework-FinderSync==8.5
pyobjc-framework-FSEvents==8.5
pyobjc-framework-GameCenter==8.5
pyobjc-framework-GameController==8.5
pyobjc-framework-GameKit==8.5
pyobjc-framework-GameplayKit==8.5
pyobjc-framework-ImageCaptureCore==8.5
pyobjc-framework-IMServicePlugIn==8.5
pyobjc-framework-InputMethodKit==8.5
pyobjc-framework-InstallerPlugins==8.5
pyobjc-framework-InstantMessage==8.5
pyobjc-framework-Intents==8.5
pyobjc-framework-IOSurface==8.5
pyobjc-framework-iTunesLibrary==8.5
pyobjc-framework-LatentSemanticMapping==8.5
pyobjc-framework-LaunchServices==8.5
pyobjc-framework-libdispatch==8.5
pyobjc-framework-LinkPresentation==8.5
pyobjc-framework-LocalAuthentication==8.5
pyobjc-framework-MapKit==8.5
pyobjc-framework-MediaAccessibility==8.5
pyobjc-framework-MediaLibrary==8.5
pyobjc-framework-MediaPlayer==8.5
pyobjc-framework-MediaToolbox==8.5
pyobjc-framework-Metal==8.5
pyobjc-framework-MetalKit==8.5
pyobjc-framework-MetalPerformanceShaders==8.5
pyobjc-framework-ModelIO==8.5
pyobjc-framework-MultipeerConnectivity==8.5
pyobjc-framework-NaturalLanguage==8.5
pyobjc-framework-NetFS==8.5
pyobjc-framework-Network==8.5
pyobjc-framework-NetworkExtension==8.5
pyobjc-framework-NotificationCenter==8.5
pyobjc-framework-OpenDirectory==8.5
pyobjc-framework-OSAKit==8.5
pyobjc-framework-OSLog==8.5
pyobjc-framework-PencilKit==8.5
pyobjc-framework-Photos==8.5
pyobjc-framework-PhotosUI==8.5
pyobjc-framework-PreferencePanes==8.5
pyobjc-framework-PushKit==8.5
pyobjc-framework-Quartz==8.5
pyobjc-framework-QuickLookThumbnailing==8.5
pyobjc-framework-SafariServices==8.5
pyobjc-framework-SceneKit==8.5
pyobjc-framework-ScreenSaver==8.5
pyobjc-framework-ScriptingBridge==8.5
pyobjc-framework-SearchKit==8.5
pyobjc-framework-Security==8.5
pyobjc-framework-SecurityFoundation==8.5
pyobjc-framework-SecurityInterface==8.5
pyobjc-framework-ServiceManagement==8.5
pyobjc-framework-Social==8.5
pyobjc-framework-SoundAnalysis==8.5
pyobjc-framework-Speech==8.5
pyobjc-framework-SpriteKit==8.5
pyobjc-framework-StoreKit==8.5
pyobjc-framework-SyncServices==8.5
pyobjc-framework-SystemConfiguration==8.5
pyobjc-framework-SystemExtensions==8.5
pyobjc-framework-UserNotifications==8.5
pyobjc-framework-VideoSubscriberAccount==8.5
pyobjc-framework-VideoToolbox==8.5
pyobjc-framework-Vision==8.5
pyobjc-framework-WebKit==8.5
pyOpenSSL==22.0.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytweening==1.0.4
rubicon-objc==0.4.2
selenium==4.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
wsproto==1.1.0
